fix(op-dispute-mon): remove unused collateral calculator
