+++ op-dispute-mon/mon/bonds/collateral.go
@@ -1,43 +1,50 @@
 package bonds
 
 import (
-	"context"
-	"fmt"
 	"math/big"
 
-	faultTypes "github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
 	monTypes "github.com/ethereum-optimism/optimism/op-dispute-mon/mon/types"
-	"github.com/ethereum-optimism/optimism/op-service/sources/batching/rpcblock"
 	"github.com/ethereum/go-ethereum/common"
-	"golang.org/x/exp/maps"
 )
 
-type BondContract interface {
-	GetCredits(ctx context.Context, block rpcblock.Block, recipients ...common.Address) ([]*big.Int, error)
+type Collateral struct {
+	// Required is the amount of collateral required to pay out bonds.
+	Required *big.Int
+
+	// Actual is the amount of collateral actually head by the DelayedWETH contract
+	Actual *big.Int
 }
 
-// CalculateRequiredCollateral determines the minimum balance required for a fault dispute game contract in order
-// to pay the outstanding bonds and credits.
-// It returns the sum of unpaid bonds from claims, plus the sum of allocated but unclaimed credits.
-func CalculateRequiredCollateral(ctx context.Context, contract BondContract, blockHash common.Hash, claims []faultTypes.Claim) (*big.Int, error) {
-	unpaidBonds := big.NewInt(0)
-	recipients := make(map[common.Address]bool)
-	for _, claim := range claims {
-		if monTypes.ResolvedBondAmount.Cmp(claim.Bond) != 0 {
-			unpaidBonds = new(big.Int).Add(unpaidBonds, claim.Bond)
+// CalculateRequiredCollateral determines the minimum balance required for each DelayedWETH contract used by a set
+// of dispute games.
+// Returns a map of DelayedWETH contract address to collateral data (required and actual amounts)
+func CalculateRequiredCollateral(games []*monTypes.EnrichedGameData) map[common.Address]Collateral {
+	result := make(map[common.Address]Collateral)
+	for _, game := range games {
+		collateral, ok := result[game.WETHContract]
+		if !ok {
+			collateral = Collateral{
+				Required: big.NewInt(0),
+				Actual:   game.ETHCollateral,
+			}
 		}
-		recipients[claim.Claimant] = true
-		if claim.CounteredBy != (common.Address{}) {
-			recipients[claim.CounteredBy] = true
+		gameRequired := requiredCollateralForGame(game)
+		collateral.Required = new(big.Int).Add(collateral.Required, gameRequired)
+		result[game.WETHContract] = collateral
+	}
+	return result
+}
+
+func requiredCollateralForGame(game *monTypes.EnrichedGameData) *big.Int {
+	required := big.NewInt(0)
+	for _, claim := range game.Claims {
+		if monTypes.ResolvedBondAmount.Cmp(claim.Bond) != 0 {
+			required = new(big.Int).Add(required, claim.Bond)
 		}
 	}
 
-	credits, err := contract.GetCredits(ctx, rpcblock.ByHash(blockHash), maps.Keys(recipients)...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load credits: %w", err)
+	for _, unclaimedCredit := range game.Credits {
+		required = new(big.Int).Add(required, unclaimedCredit)
 	}
-	for _, credit := range credits {
-		unpaidBonds = new(big.Int).Add(unpaidBonds, credit)
-	}
-	return unpaidBonds, nil
+	return required
 }
